Meeting the requirements about classes and objects

+++ Prática 2 - RMI/database.py
@@ -4,16 +4,18 @@ import produto
 class DB:
 
     def __init__(self):
-        self.conexao = sqlite3.connect("database.db")
-        cursor = self.conexao.cursor()
+        conexao = sqlite3.connect("database.db")
+        cursor = conexao.cursor()
         cursor.execute("CREATE TABLE IF NOT EXISTS mercado(codbar INTEGER PRIMARY KEY, nome, estoque, loc, preco)")
-        self.conexao.commit()
+        conexao.commit()
         cursor.close()
+        conexao.close()
 
 
     def inserir(self, produto):
         try:
-            cursor = self.conexao.cursor()
+            conexao = sqlite3.connect("database.db")
+            cursor = conexao.cursor()
             cursor.execute("INSERT INTO mercado(codbar, nome, estoque, loc, preco) VALUES (?, ?, ?, ?, ?)", (produto.codbar, produto.nome, produto.estoque, produto.loc, produto.preco))
 
             if (cursor.rowcount > 0):
@@ -21,8 +23,9 @@ class DB:
             else:
                 id = None
 
-            self.conexao.commit()
+            conexao.commit()
             cursor.close()
+            conexao.close()
             return id
         except sqlite3.IntegrityError:
             return -1
@@ -32,13 +35,15 @@ class DB:
     
 
     def buscar(self, codbar):
-        cursor = self.conexao.cursor()
+        conexao = sqlite3.connect("database.db")
+        cursor = conexao.cursor()
         cursor.execute("SELECT * FROM mercado WHERE codbar = ?", (codbar,))
         retorno = cursor.fetchone()
         cursor.close()
+        conexao.close()
         
         if (retorno == None):
-            return None
+            return retorno
         else:
             produto_encontrado = produto.Produto(retorno[0], retorno[1], retorno[2], retorno[3], retorno[4])
             return produto_encontrado
@@ -46,20 +51,30 @@ class DB:
     
 
     def atualizar(self, produto):
-        cursor = self.conexao.cursor()
+        conexao = sqlite3.connect("database.db")
+        cursor = conexao.cursor()
         cursor.execute("UPDATE mercado SET nome = ?, estoque = ?, loc = ?, preco = ? WHERE codbar = ?", (produto.nome, produto.estoque, produto.loc, produto.preco, produto.codbar))
-        self.conexao.commit()
+        conexao.commit()
 
         conf = cursor.rowcount > 0
         cursor.close()
+        conexao.close()
         return conf
 
 
     def deletar(self, id):
-        cursor = self.conexao.cursor()
-        cursor.execute("DELETE FROM mercado WHERE codbar = ?", (id,))
-        self.conexao.commit()
+        conexao = sqlite3.connect("database.db")
+        cursor = conexao.cursor()
         cursor.execute("SELECT * FROM mercado WHERE codbar = ?", (id,))
         retorno = cursor.fetchone()
-        cursor.close()
-        return retorno
+        if (retorno == None):
+            cursor.close()
+            conexao.close()
+            return retorno
+        else:
+            cursor.execute("DELETE FROM mercado WHERE codbar = ?", (id,))
+            conexao.commit()
+            produto_encontrado = produto.Produto(retorno[0], retorno[1], retorno[2], retorno[3], retorno[4])
+            cursor.close()
+            conexao.close()
+            return produto_encontrado
